Updated FerociousBlade recipe to be harder and moved it to Legendary package
